Disable search field configuration until ready for impelmentation
We have all of the admin UI build out for configuring search fields;
however, the Blacklight config for search is slightly more complex
than it is for facet and view configurations.

Since we want to be able to show facet and view configuration features
without the distraction of having to explain that the search
features are not working right now, we're disabling the UI for
search config for the time being.