Update solr host and core dynamically
ISSUE
If an admin creates a new active config with a different Solr host
or core while the application is currently running, we need to re-
set those values in the current blacklight controller.

FIX
Add code to reset the host and core whenever we update the active
Blacklight configuration.