Update to Bootstrap 5.3.0 (for color mode support)
Going forward, Bootstrap is moving toward extracting color
settings to make applying themed color schemes easier.
https://getbootstrap.com/docs/5.3/customize/color-modes/

In order to align our themeing and branding work with this
change, we're updating to the latest alpha release of
the Bootstrap gem, so that it's easier to align our color
management with Bootstrap's direction.