Implement local HTML layout file
In reviewing other Blackligh implementations like Stanford's
SearchWorks, it appears that folks are just making a local copy
of the base Blacklight layout file if they need to modify top-level
layout details such as header tags.

Adopting this strategy allows us to inlcude the dynamic stylesheet
in the layout rather than having to patch it in with JavaScript
after the initiall stylesheets are loaded.