Allow admins to upload a unique logo for each Theme
We want the ability to set a logo on a per-theme basis since
folks may eventually be updating their branding or logos and want
to preview the new artwork before going live to public views.

I.E. we want to be able to keep an 'old' logo associated with the
active theme while experimenting with new logo versions in previews.

This change also introduces Rails ActiveStorage to manage file uploads and
attachments.