Add theme previews
This change gives the admin a 'thumbnail' preview of each available
theme to allow quick visual scanning and selection of themes.

The thumbnail is produced by rendering the catalog homepage in an
iframe element.  The strategy uses here borrows heavily from this post:
https://medium.com/@jamesfuthey/simulating-the-creation-of-website-thumbnail-screenshots-using-iframes-7145269891db

There's a fair bit of other CSS added to make the previews layout
nicely in the dashboard.  We've implemented a loading spinner that
uses pure-css to animate a loading graphic.  The loaded iframe
displays over the loading icon and obscures it.  The animation
is currently set to run for 30 cycles (~90 seconds) and then stop.
This should generally be more than enough for general load cases.

We've also added support to dynamically preview a theme by adding
the query parameter `?theme_id=[:id]` to any request.