Add preliminary color support for search bar
This initial pass employs two different methods to inject dynamic
CSS (and remove undesired Blacklight styling):
* Redefine Blacklight view components directly
* Define new view components and dynamically replace the provided
  Blacklight view components

This provides us with a working proof-of-concept, but eventually
we need a consistent pattern.

View components are intended to allow us to write simple replacements,
but so far it feels like we've had to bring in a large amount of code
to make relatively minor styling changes.

Ideally, we'd like to be able to make smaller interventions into
Blacklight.  It's not clear whether the current state is just how
view components work or whether it's a side effect of Blacklight
incorporating too much styling into the components directly.

TODO:
* Choose one of the two patterns: override vs. inject new components
* Figure out if there's way to achieve the styling results without
  needing to touch so many components