Update application to use encrypted credentials
This change ensures that a private master.key file is persisted
in the config directory across deployments made via Capistrano.

Rather than maintaining a separate secret_key_base in screts.yml,
the value is now included in the encrypted credentials file.

The master.key value is stored outside the repository in a password
management application and manually distrubuted to environments that
require it.