Set "provider" to "local" for database users
We want to show the "account type" in the user index.  We can use the
provider name, e.g. 'google', for OmniAuth accounts, but database
accounts previously had the provider set to `nil`.

Rather than writing a helper that checks for nil every time we want
to display the user's account type, i.e. nil => 'local', we chose to
set the value once when the user is initialized.  Now we can just
display the provider value directly.