Refactor dynamic field configuration
This commit moves the dynamic mapping of catalog fields from
the CatalogController to the Config class. This reduces complexity
in the CatalogController and moves the mapping into the persistence
object where it fits better. (e.g. no more obsessive calls to
`Config.current...`)

A group of Config methods that have become obsolete through the
evolution of the `update_catalog_controller` have also been removed
in this change.