Add a wrapper for interacting with certbot
We want to be able to update the TLS certificate for the application
website so the repository owner can access the site via a custom
domain name.

This code provides a wrapper for the certbot command line interface
to LetsEncrypt.  It makes a large number of simplifying assumptions
about certificate interactions based on our required user flow.

This wrapper does not attempt to provide a general purpose interface
to the LetsEncrypt service and only provides access to a very limited
set of certbot features.

Unsing the wrapper, the application can
* Read the current certificate info including domains and expiration
* Add or Remove a single domain to the certificate at a time