Remove dead code from the Custom Domain controller
We do not provide a way to view individual domains, all the relevant
information is available on the index view.  So there is no route
to a show action.

The cancancan load_and_authorize_resource handles setting the domain
for singular actions, so we also do not need the set_custom_domain
method.