Add logging for Certbot calls
For debugging purposes (info), we want to be able to see read-only
calls to Certbot.

For auditing purposes (warn), we want to log all calls to Certbot that
might make modifications to a certificate.

This information helps us implement the least-privelege necessary to
allow the rails user (deploy) to update certificates.