Refactor CustomDomain tests to use Certbot TestClient
**ISSUE**
The code previously stubbed out low-level system calls to Open3
to simulate behavior in the Certbot::V2::Client that triggered
various behaviors in the CustomDomain model.  This requires
the tests to have knowlege of the underlying implementation in
a different class.  This makes the tests somewhat fragile and
difficult to interpret.

**FIX**
This change leverages the Certbot::V2::TestClient to directly set
the return values that we want to test in the CustomDomain model.
This decouples the CustomDomain tests from the Certbot implementation
and hopefully clarifies the specific return values and object states
that drive behavior in the CustomDomain class.