Provide detailed error instructions for CustomDomain host errors
**ISSUE**
When a custom domain setup error occurs, we want to provide the
admiistrator with enough information to correct the error.  Since
admins may not be familiar with setting up TLS certificates, a
basic Rails error message may not be sufficient to let them know
what has gone wrong.

**SOLUTION**
Suppress the default error messages for :host errors and provide
more detailed setup instructions based on the error type.

**NOTE**
Curently the prioritized list of errors is:

* :certbot      - There was a problem with the Certbot installation
* :format       - The domain name is missing or improperly formatted
* :taken        - The domain name is already in use by the application
* :unresolvable - The domain name can not be resolved via DNS
* :certificate  - Certbot encountered an error when trying to add the domain