Disable user self-registration
**ISSUE**
Users should not be able to create accounts on their own without
being invitied by an administrator.

**SOLUTION**
Remove the Devise Registerable module from the user model.  This
removes the "Sign Up" link from the login screen and removes Devise
funcitonality to allow users to edit and delete their own account.

Because Blacklight inlcudes a link to the edit_user_registration
route, we need to supply a definition for this helper.  For the time
being, we've added a user profile route that allows a user to
edit their own record.