Update the application to manage a single configuration
**ISSUE**
There is no requirement to support multiple configurations that
administators can choose between.  The current code always used
the last created config with no ability to activate an earlier config.

This lead to more complex code and more complex UI than users
require.

**SOLUTION**
Implement a modified singleton pattern for the configuration data
(which is basically global state).  This allows us to make a number
of simplifying assumptions:
* There will always be a single configuration persisted to the database
* Users do not require creation or selection workflows for the
  configuration
* We can pre-populate the applicaiton with sane defaults for Solr

As a side benefit, this means that new production instances do not
trigger Catalog / Solr errors on startup because they now have a
valid bootstrap configuration.