Redirect admins to the dashboard after login
**STORY**
As an administrator, I want to go directly to the dashboard after
logging in, so that I don't have to always click the dashboard link.

**DETAILS**
This change updates the Devise `.after_sign_in_path_for(resource)`
method to check for admin users and redirect them to the dashboard
instead of the Devise default.

Normal users are still sent to the root_path, which is the Blacklight
Catalog index page.

**NOTES**
Because the catalog controller queries Solr to provide data rendered
in the index view, we need to stub the Solr service to return the
minimally expected empty response.