Fix host url display on custom domains page
**ISSUE**
Only the hostname portion of the default URL is displaying on
the Custom Domains index view.  We want the full domain to display
so the URL is actually resolvable

**FIX**
Supply the url to the link helper twice, as both the link reference
and the display text to be used for the link.