Allow admins to set display name and roles for invited users
**STORY**
As an administrator, I would like a one-step process to set a
user's display name and roles when I invite them, so that I don't
have to remember to edit the user after inviting them.

**SOLUTION**
Add optional fields to the invitation view to set the display
name and roles.  These can also be left blank and modified later by
editing the user.