Allow administrators to deactivate users
Instead of deleting users from the database, allow administrators
to deactivate an account to prevent the user from logging in, while
retaining user data.  We want to keep user references intact so that
we can audit objects and actions even after the user had been
disabled in the system.

This change:
* Allows user records to be marked as deativated
* Adds UI functionality to allow authorized users to
  deactivate other users
* Adds 'deactivated' to the possible user states displayed in
  the user dashboard
* Adds the user account status to the User show view