Fix for admins not redirecting to dashboard on login
**ISSUE**
Administrators logging in with Google (OmniAuth) accounts were not
being redirected to the dashboard on login.

**SOLUTION**
Remove the custom `after_sign_in_path_for` method in the
OnmiauthCallbacksController that was preventing the newer version
of this method in the ApplicationController from being called.

Now both OmniAuth and Database users call the same method.