Refactor UserHelper#duration to reduce complexity
Instead of using a rescue clause, add a guard clause that ensures
the expected data behavior.

In this case, we want to ensure that the timestamp value can be
converted to a Time value.  Calling `.try` ensures we get a usable
time value from strings.  I.E. strings without a valid time value
will return nil and the guard clause will take effect.