Read hostname from system
**ISSUE**
As a simplifying assumption in early-stage development, we hardcoded
the server hostname "t3-dev.curationexperts.com" into the mailer
configuration.  We now want to run multiple environemnts and the static
hostname is incorrect on other servers.

**FIX**
In the initializer, read the fully qualified name configured on the
system instead of using a hardcoded value.