Add basic UI and model for Blueprints
Blueprints are intended to contain all of the configuration information
such as field definitions, search behaviors, translations, and other
settings required to define the object classes managed in the
repository.

We've chosen the term 'blueprint' since model, class, and other related
terms are generally already overloaded in both the software and library
domains.  Further, borrowing from the idea of an architectural blueprint,
this type of plan can contain multiple different layers or types of
schematics (structural, electrical, hvac) that combine to completely
describe a finished real-world instance.  In a similar way, a
blueprint here can define indexing, search, input constraints, etc. that
govern unique work instances in the repository.