Improve UI for Blueprints with ability to manage fields
Replace the generated Blueprint views with custom views that support
a cleaner and more streamlined admin experience.

Add the ability to manage individual fiend definitions for a blueprint.

TECHNICAL NOTE:
Fields are stored as a JSON (jsonb) blob in the database.  This
simplifies management of field ordering (i.e. we don't need a separate
order field or data structure) and ensures that fields' lifecycles
are managed in parallel with their respective blueprint (i.e. this
architecture prevents the possiblity of orphan fields).  It also
avoids complications that would occur if a field were shared between
two blueprints, but edited in only one.