Implement backround processing for Ingest batches
When an administrator uploads a new manifest for ingest, submit a
background job to iterate through each record in the manifest.

The current implementation does not do anything with the record, but
adds a process counter and status updates to display whether the
Ingest is 'queued', 'processing', 'completed', or 'errored'.

TODO
* Implement a version of the #process_record method in the ImportJob
class that reads the supplied metadata and creates the corresponding
repository object.
* Update the Administrator UI to display the processing status