Add an animated progress bar to ingests
This change adds a helper and CSS that formats Ingest information
in a badge-like component.

The text in the badge provides the basic processing status of the
Ingest including the number of items processed.  CSS is used to
add color cues for the status along with animation showing how
many items in the corresponding job have been processed so far.