Refactor Config records to use shared Fields
**ISSUE**
In order to synchronize configuration between Blueprint objects,
the CatalogController, and the core application Config record, we
want to reference a single set of Field objects stored in the database
rather than trying to synchronize changes to JSON hashes stored
separately in the Config and Blueprint records.

**ACTION**
Replace JSON FieldConfig objects used by the Config record with
Field records.  This has the added benfit of giving us a single
configuration point for Fields so we don't have to manage nested
forms in the Config views.

**NOTE**
As part of this change, we have refine some of our naming, in
particlar:
* We are now using 'name' instead of 'display_label' for the
corresponding attribute and class names in anticipation of
adding internationalization features at a later date to provide
multi-lingual display labels.
* We are now using the wording 'list_view' instead of 'search_results'
to indicate when a field should show up in search results.  This
provides logical parity to 'item_view'.
* We've realized that 'solr_field_name' is long and somewhat redundant,
so we shortened it to 'solr_field'