Remove item from Solr when destroying Items
**ISSUE**
Blacklight uses the item metadata stored in Solr to display the record.
So, even if we destroyed an item in the database, it would still
show up in searches and item views if we had indexed it into solr.

**SOLUTION**
Add a callback that deletes the corresponding ID from Solr indexes
when we remove an item from the database.  We remove the indexed
document only after successfully commiting the deletion to the database
so we don't end up with unfindable objects if the deletion fails
for some reason.

**NOTE**
This commit also streamlines Solr calls by passing a commit flag on
add and delete updates rather than making a second Solr call to
trigger the commit.