Revert "Fix flaky timeout in certificate test"
This reverts commit 2be2d58d8930ed848cc0cf1cfcba73e1ba1d2da5.

The new test doesn't actually cover the needed rescue clause and
I can no longer replicate the timeout issue that inspired this change
so we're reverting back to the earlier test setup which seems to be
working without issue.  It seems like there may have been some strange
state stuck in the DNS cache or networking state that's now cleared.