Update CatalogController configuration when Fields are modified
**ISSUE**
Changes were not appearing in the UI when we made changes to Fields
that should effect the CatalogController behavior.  E.G. if we changed
a field to display it in the search result view, we needed to
restart the application or manually reload the Blacklight field
configurations for the change to appear to users.

**SOLUTION**
Update the CatalogController field configurations any time a field
is modified.  Some field changes may not require a change to the
CatalogController, but we've chosen to err on the side of never missing
a relevant change.

In the future, we may want to more narrowly define the update process
so it only changes config for the field being saved rather than
reloading the config for all active fields.