Remove dead code
Remove the code used to infer fields from an existing Solr index.

The current use cases don't include attaching to an existing Solr
core.  Instead, we'll re-introduce similar functionality to import
from a legacy Solr core as an add-on in the future.  For now, the
code provides no value and adds cognitive complexity to the Config
model.