Remove duplicated code between Field and Config models
Both the Field and Config model need to update the Blacklight
configuration at certain times.  For expediency we initially
largly duplicated the code in each class.

This change refactors the Field model to rely on the catalog
update code in the Config model to eliminate the duplication and
make future maintenance easier.