Add routing and request handling for updating Field sequence
This change adds a RESTful route for the `move` action on fields.

Note that the move action typically makes changes to other fields
in additon to the field it is called on. I.E. most moves will update
the sequence number for the subject field and one or more other fields.

* `up` decrement the field's sequence order (i.e. move it closer to
the head of the list)
* `down` increment the field's sequence order (i.e. move it closer to
the tail of the list)
* `top` move the field to the head of the list
* `bottom` move the field to the tail of the list

Calling `up` or `top` when the field is already at the top of the list
will leave sequence numbers unchanged.

Calling `down` or `botton` when the field is already at the tail of
the list will will the sequence numbers unchanged.