Allow admins to re-order fields
This commit adds user interface controls in the Fields index view
to allow system administrators to modify the order fields are
displayed in.

The field order is used in multiple locations in the application:
* Field dashboard
* Blueprint dashboard - viewing fields in an individual blueprint
* Search result listings
* Item view pages

A single field order is used across the application to support a
user's ability to scan content across the application.