Add Ingest reporting
**STORY**
As an administrator, I want to be able to drill down into the status
of my Ingest jobs, especially if something went wrong.

**DETAILS**
This commit adds ingest reporting and persists the report for
administrators to review later.  Simple item-level status is tracked.
If any errors occur, the error type and message are captured and
basic context is recorded to help diganose and debug any issues.