Display the number of errors in an Ingest
This change adds a field to persist the number of errors encountered
while processing an ImportJob.  This allows us to display top-level
context about the Ingest without having to parse the associated
report file.

If a user is interested in drillling down for details, they can
download the Ingest report and review it manually for more detail.