Use the first active ordered field as the Blacklight title field
Blacklight requries a title field to be configured for the application.
This field is used as the linkable title for documents in search
results and as the item header in single-item views.

Previously, we set this by looking for the first field that matched the
string "Title".  Over time we found that
1. We wanted to use the word "Title" in the name of more than one field
   e.g. "Alternate Titles" and "Primary Title".
2. We typically ordered fields so title was at the top of the list to
   reflect how blacklight displays field ordering.

With these insights, it became clear that using the first field in the
ordered field list would be the easiest solution for users to manage.
It also has the benefit of being clear and easy to implement from a code
perspective.