Output semantic metadata to logs in JSON format
In order to easily leaverage search and filtering capabilities in
log aggregation and analysis tools like AWS CloudWatch logs, we want
to send our logs in an easily parsable format.

We've chosen [Semantic Logger](https://logger.rocketjob.io/index.html)
because it provides an easy drop-in replacement for the default Rails
logger. Semantic Logger doesn't require us to change any of our existing
logging code and makes it simple to output logs in JSON format.

We also take advantage of the [metric](https://logger.rocketjob.io/metrics.html)
features of Semantic Logger to capture performance (e.g. duration)
statistics for areas of the code that we want to monitor for
potential bottlenecks as the codebase grows and evolves.