Move button for new Ingests to top of form
**ISSUE**
When the list of ingests gets longer than one screen, it becomes
less convenient to scroll to the end of the list to click the button
to submit a new job. Some users might even miss the button because
the didn't think to scroll to the bottom.

**RESOLUTION**
Move the button to start a new ingest to the top of the screen so
it is always visible when the screen first renders.