Validate required fields when saving Items
This commit adds validations that check each field marked as
required and ensures that the field is not blank for the item.

If the field is blank, the validation adds an error indicating that
the field name can not be blank.

This change also adds a check to ensure that description is not blank.
Item tests and factories are updated to reflect that description
is no longer valid when blank.