Rename Item 'description' to 'metadata' (⛔️ Requires Downtime)
After some time using the name 'description' to refer to the
JSON metadata field that describes an item, the name has come
to seem a bit awkward in general use.  It also feels clunky when
referring to a client field named 'description': we get constructs
like `description['description']` or `description.description`

This change modifies those usages to `metadata['description']`
and `metadata.description`.

More generally, referring to an item's metadata feels more consistent
with the library domain where description is just one subset of
metadata.

⛔️ DOWNTIME EXPLANATION
The table migration will invalidate the database for commits pior
to this change becasuse the old code will refere to an attribute
name `description` while the database will have a column `metadata`.

To address this, stop any running instances of the application
before deploying, allow the migration and code deployment to be
completed successfully with no live instances accessing the database.
Once the column rename in the migration is complete, the application
can be restarted with the updated code and the new attribute name
will map to the correct database column.