Store Solr download files in persistent directory
**ISSUE**
In development environments, `solr_wrapper` downloads files to a temp
folder under `/var/folders`.  On OSX, these folders are deleted after
each reboot and the solr executables need to be redownloaded.  This
can be frustrating when working on a slow network.

**SOLUTION**
Configure `solr_wrapper` to download the executable archives to a folder
that persists across reboots.