Add UI elements to manage Collections
This change
* Adds the route, controller, and views necessary to manage Collections
  from the browser
* Adds a navigation item in the admin dashboard for Collections

NOTES:
Collections and Items intentionally share most behavior so that
administrators don't have to learn or remember different user
interfaces for similar functionality in different parts
of the application.

However, because we always anticipate the total collection count to be an
order of magnitude or more smaller than the number of items in the
repository (i.e. collections often number in the 10s while even a
small respository will likely have thousands of items), we prefer
to store collection data in a separate table from items.