Fix ForbiddenAttributesError when submitting collections
**ISSUE**
Creating a new collection or attempting to edit an existing
existing collection raises a `ForbiddenAttributesError` exception
when you submit the form.

**DIAGNOSIS**
The form inherited from the Items controller is submitting paramaters
as
```
'collection' => { 'blueprint' => ..., 'metadata' => { ...}}
```
but the controller only permits prams scoped under `items`, i.e.
```
'items' => { 'blueprint' => ..., 'metadata' => { ...}}
```

**RESOLUTION**
Add an explict scope to the form to override the default scoping
inferred from the model being passed to the form.