Add a custom form helper for displaying collection options
**ISSUE**
We want to be able to provide a collection drop-down selection list
in our collection and item edit forms, but we want it to handle
multivalue fields by displaying individual inputs that aggregate
as an array valued parameter - i.e. each takes a single value and the
field name ends in '[]' - instead of a multi-value select box.

**SOLUTION**
Implement a custom input that provides the desired behaviors for our
forms.