Add collection selection UI to edit forms
This change adds the code necessary to allow administrators to
set a field's data_type to `vocabulary` and have the edit form
display a list of availabel collections in a selection dropdown.

Right now, the `vocabulary` data_type only supports collections,
but we hope to generalize the funcionality to other controlled
vocabularies in the future.