Refactor Item and Collection model tests to use shared examples
The tests for the Item model and the Collection model had diverged
as we more fully implemented Collections.  We also discovered cases
where we had hard-coded Item class operations into the Resource class.

This commit merges the Item and Collection tests and refactors them
to be run as [RSpec shared examples](https://rspec.info/features/3-12/rspec-core/example-groups/shared-examples/).

It also implements minor corrections to the Resource class to
ensure class methods work with the desired subclasses.