Fix bug that allows Resources to be saved without required fields
**ISSUE**
The Resource edit form (i.e. Items & Collections) would allow
administrators to save resources without filling in required fields.

**DIAGNOSIS**
We were initially checking for the presence of the required Field's
key rather than its value. In addition, for multi-value fields, we
need to collapse empty input like ['', ''] that can be returned from
the form.

**FIX**
Update the validation test with more robust examples of 'blank' values
and update the model code to check for the various possible permutations
of an empty field.