[REFACTOR] Handle resource labels for missing titles
It "shouldn't" happen in the wild, but there are situations in
development and testing where we may have a resource with an
empty title field (i.e. first field defined in the blueprint).

This refactoring ensures that #label always returns some identifiable
label, either the "Title" defined by the blueprint, or the
resource class and ID as a placeholder.

We've also memoized the label's field name once a blueprint is loaded
since we don't want to keep having to reload Fields each time we want
to look up the label value.