Fix flaky request test
Depending on the test order, there may not be a current Theme defined
yet, casuing any request to create one as part of the application's
layout and CSS rendering process.

If tests that test Theme counts are run without a current theme defined,
a new current theme will be created. This typically throws off the
expectect count in the test.