[REFACTOR] Force dev server to start on port 3000
Foreman auto increments the PORT environment variable for each line
in the Procfile. This means the rails server port becomes position
dependent - i.e. putting the webserver on line 2 causes the service
to launch on port 3100.

We want to start Solr first, so it has time to initialize before we
start our web server. So the webserver comes on line 2 of the
Procfile (which sets the environment PORT=3100). We want the server
to run on port 3000, so we need to set the value explicitly instead
of reading the PORT variable from the environment.