Speed up screen refresh when reordering fields
This change uses turbo frames to update only the fields
table when re-ordering fields instead of re-drawing the
entire page. This results in a slightly faster refresh.

Turbo retains the broswer scroll state so we don't scroll back
to the top of the page when re-ordering fields at the end of
the list. This helps provide a less distracting user experience
than having the page unexpectedly jump away from the fields
the user is viewing.