Fix flaky search_bar_component coverage
**ISSUE**
In a small subset of random test seeds, the search_bar_component
code would not be tested, resulting in less than 100% test coverage.

**DIAGNOSIS**
One of the Field model tests was resetting the CatalogController
blacklight configuration to an empty configuration.  This config
includes a pointer to the header component:
https://github.com/curationexperts/t3/blob/v0.24.0/app/controllers/catalog_controller.rb#L27
If this configuration is missing, the header component and the
search bar component are never called.

**FIX**
Stub the blacklight configuration only for the example group
instead of setting it to a new blank configuration that is used in
all subsequent tests.