[REFACTOR] Clarify ability determination
**ISSUE**
As we've been adding additional classes that can be managed in the
dashboard, the mapping of role names to managed classes has gotten
somewhat longer. It's a little difficult to scan this larger list
to understand exactly which roles manage which classes.

This update inverts the mapping to make it more obvious which
classes map to each specific role.