Display Field configuration in JSON format
**STORY**
We want to be able to easily export the Field configuration from one
repository instance to another, so that we can quickly configure
a new environment to match an existing one.

**SOLUTION**
Allow Field configuration to be exported in JSON format, so that we
can later use it to configure a different repository instance.