Fix file download bug in production
**ISSUE**
When we reference a HTTP or HTTPS URI for a file in an import manifest,
we're seeing the following error in production:
```
private method `open' called for #<URI::HTTPS https:// ... >
```

**DIAGNOSIS**
The 'open-uri' module appears to be loaded for background workers in
the defelopment environment automatically, but not for background
jobs in the production environment.

**FIX**
Explicitly require the 'open-uri' module in the job class where
the error is occurring.