[REFACTOR] Rename Config --> SolrService
The Config model is primarily concerned with managing the connection
to a Solr service. We want to add functionality to import and export
the system "configuration" which more broadly includes all
administrative settings. To avoid conceptual and naming conflicts going
forward, we're renaming the class to better match it's specific
use and functionality.