[REFACTOR] Cache the Field order
**ISSUE**
Many places in the code require a list of Fields in their sequence
order. Previously, the Fields were loaded from the database each
time this list was needed. The list is *very* stable in a production
environment, only changing when fields are added or re-ordered, which
occurrs very rarely.

**RESOLUTION**
Implement a class method that caches the Fields in order. The only events
that could invalidate the order are saves to the database. So, by
reseeting the cached list on any save, we save a significant number
of database read operations.