Fix flaky SolrService tests
**ISSUE**
If the test suite crashes and/or in some randomized sequences,
the test suite was attempting to create SolrService records when
a record already existed in the database. Because of the singleton
validations, this would cause some tests to fail.

**REOLUTION**
Refactor the tests to use the current instance where appropriate;
otherwise, modify the tests as needed to ensure they respect the
singleton validations.