Add User Interface for managing terms
The interface comes largely from scaffold generators, so it's
functional but perhaps not pretty.

Basic vocabulary and term management functions are supported
including creating, editing, and deleting of vocabularies and the
terms within them.

Terms are always managed within the context of a specific vocabulary.
This means that the same term can exist within different vocabularies,
but a single vocabulary may not contain duplicate terms.

URLs for vocabularies and terms use human-friendly slugs instead of
opaque identifiers to identify objects.

This means that URLs for terms always include the corresponding
vocabulary slug in their path.  e.g.

```
/vocabularies/resource-types/term/thesis-or-dissertation
/vocabularies/submission-class/terms/thesis-or-dissertation
```

Under-the-hood, objects still use traditional numeric identifiers,
but all route processing is based on the human-friendly slug.