Allow administrators to import Vocabulary Terms from a JSON file
**STORY**
As an administrator, I want to be able to quickly populate default
vocabulary settings and terms, so that I don't have to manually
enter each vocabulary term via the user interface.

**SOLUTION**
Allow the import of vocabulary and term data from a JSON file with
the same structure as the config file export format.