[REFACTOR] Standardize Term attribute names
**RATIONALE**
We want to use consistent attribute naming across models to make
code maintenance easier. In this case, we had used `:slug` for the
semantic identifier that gets used in URLs. This word gets used in
SEO contexts, but isn't as meaningful in library and archive contexts.

The pattern going forward is:
* `:label` The human-readable title of a thing. This can include spaces and special characters
* `:key` The url-friendly identifier that can contain only as subset of printable ASCII characters
* `:note` A description, usage note, or other text that provides context for the record

**ALSO SEE**
Pull Request ##258