[REFACTOR] Remove unused code
**RATIONALE**
Coverage reports indicate that this code is no longer being tested
when we run the test suite. This is a side effect of the fact that
we no longer rely on inferred partials in vocabulary views.

For now, we're remvoing the related code to avoid having to
separately test and maintain it. If the view architecture changes
at some point in the future, we can easily replace it.